zmiana obliczeń ECDSA na całkowitoliczbowe

+++ digital-signature/src/main/java/pl/mkoi/project/signs/EcdsaSign.java
@@ -8,19 +8,19 @@ public class EcdsaSign implements Serializable {
 
   private static final long serialVersionUID = 1L;
   
-  private BigDecimal factorR;
+  private BigInteger factorR;
   private BigInteger factorS;
 
-  public EcdsaSign(BigDecimal factorR, BigInteger factorS) {
+  public EcdsaSign(BigInteger factorR, BigInteger factorS) {
     this.factorR = factorR;
     this.factorS = factorS;
   }
 
-  public BigDecimal getFactorR() {
+  public BigInteger getFactorR() {
     return factorR;
   }
 
-  public void setFactorR(BigDecimal factorR) {
+  public void setFactorR(BigInteger factorR) {
     this.factorR = factorR;
   }
